requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
pysentimiento
flask
gunicorn
torch==1.10.1+cpu
tweepy
python-decouple
matplotlib
wordcloud
nltk
chatterbot
chatterbot-corpus